+++ src/Element/InputInterface.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * UF Form Generator
+ *
+ * @link      https://github.com/lcharette/UF_FormGenerator
+ * @copyright Copyright (c) 2017 Louis Charette
+ * @license   https://github.com/lcharette/UF_FormGenerator/blob/master/LICENSE (MIT License)
+ */
+namespace UserFrosting\Sprinkle\FormGenerator\Element;
+
+/**
+ * InputInterface
+ *
+ * Interface for Form elements classes
+ */
+interface InputInterface {
+    public function __construct($name, $element, $value = null);
+    public function parse();
+}